requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
alembic
passlib[bcrypt]
python-dotenv
python-jose